+++ dae/tests/integration/study_query_variants/test_best_state.py
@@ -44,7 +44,7 @@ f2       c2       d2    m2    2    2     prb
 @pytest.fixture(scope="module")
 def imported_study(tmp_path_factory, best_state_vcf, genotype_storage):
     # pylint: disable=import-outside-toplevel
-    from ...foobar_import import foobar_gpf
+    from dae.testing.foobar_import import foobar_gpf
 
     root_path = tmp_path_factory.mktemp(genotype_storage.storage_id)
     gpf_instance = foobar_gpf(root_path, genotype_storage)
