+++ dae/tests/integration/study_query_variants/test_foobar_vcf_queries.py
@@ -4,7 +4,7 @@ import pytest
 from dae.utils.regions import Region
 from dae.testing import setup_pedigree, setup_vcf, vcf_study
 
-from ...foobar_import import foobar_gpf
+from dae.testing.foobar_import import foobar_gpf
 
 
 @pytest.fixture(scope="module")
