+++ dae/tests/integration/tools/test_flexible_denovo_import.py
@@ -1,7 +1,7 @@
 # pylint: disable=W0621,C0114,C0116,W0212,W0613
 import pytest
 
-from tests.foobar_import import foobar_gpf
+from dae.testing.foobar_import import foobar_gpf
 from dae.tools.simple_study_import import main
 from dae.gpf_instance import GPFInstance
